Fix metadata exports on client pages for Next.js build.
Split orderbook, quote-inspector, and settings into server page wrappers
with metadata plus dedicated client components, matching the swap page pattern.
